Cat JS added...

+++ app/Http/Controllers/WebhookController.php
@@ -12,23 +12,26 @@ class WebhookController extends Controller
     public function mark_online(Request $request)
     {
         $req = json_decode($request->getContent());
-        $user=User::find($req->events[0]->user_id);
+        $user = User::find($req->events[0]->user_id);
         if ($req->events[0]->name == "member_added") {
-            $oUser = new OnlineUser();
-            $oUser->user_id = $req->events[0]->user_id;
-            $oUser->is_expert=$user->hasRole('expert');
-            $oUser->save();
-        } else {
-            if($user->hasRole('user')){
-                $ue=UserHasExpert::where(['user_id'=>$user->id])->first();
-                $oe=OnlineUser::where(['user_id'=>$ue->expert_id,'is_busy'=>true])->get()->first();
-                if(!empty($oe)){
-                    $oe->is_busy=false;
-                    $oe->save();
-                    $ue->delete();
-                }
-                OnlineUser::where('user_id',$req->events[0]->user_id)->first()->delete();
+            OnlineUser::where('user_id', $req->events[0]->user_id)->delete();
+            if (!$user->hasRole('expert')) {
+                $oUser = new OnlineUser();
+                $oUser->user_id = $req->events[0]->user_id;
+                $oUser->is_expert = 0;
+                $oUser->save();
             }
+        } else {
+            if ($user->hasRole('user')) {
+                $ue = UserHasExpert::where(['user_id' => $user->id])->first();
+                $oe = OnlineUser::where(['user_id' => $ue->expert_id, 'is_busy' => true])->get()->first();
+                if (!empty($oe)) {
+                    $oe->is_busy = false;
+                    $oe->save();
+                    (!empty($ue)) && $ue->delete();
+                }
+            }
+            OnlineUser::where('user_id', $req->events[0]->user_id)->delete();
         }
     }
 }
